requirements.txt
channels==4.0.0
channels-redis==4.0.0
crispy-bootstrap5==0.7
daphne==4.0.0
Django==4.1.4
django-crispy-forms==1.14.0
redis==4.4.0
requests==2.28.1
uvicorn==0.20.0
ptvsd==4.3.2
django-extensions==3.2.3
python-decouple==3.8